+++ plone/app/event/vocabularies.py
@@ -9,7 +9,6 @@ from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
-import pytz
 import random